Finalise locus-based multi-threading
