+++ src/tinypeel/peeling/peeling_kernels.py
@@ -1,0 +1,517 @@
+"""Low-level numerical kernels for peeling."""
+
+from numba import jit, float32
+import numpy as np
+
+
+@jit(nopython=True, nogil=True)
+def get_joint_parents(prob_sire, prob_dam, n_loci):
+    """Creates the joint parental genotypes based on the probabilities for each parent.
+
+    :param prob_sire: the probability of each genotype of each locus of the sire
+        with information of the current child from previous peeling cycle (P(p))
+    :type prob_sire: 2D numpy array of float32 with size 4 x n_loci
+    :param prob_dam: the probability of each genotype of each locus of the dam
+        with information of the current child from previous peeling cycle (P(m))
+    :type prob_dam: 2D numpy array of float32 with size 4 x n_loci
+    :return: the probabilities of all the combinations of the sire's genotype and the dam's genotype
+        with information from previous peeling cycle (P(p, m))
+    :rtype: 3D numpy array of float32 with size 4 x 4 x n_loci
+    """
+    # Equivalent einsum: output = np.einsum("ai, bi -> abi", prob_sire, prob_dam)
+    output = np.full(shape=(4, 4, n_loci), fill_value=0, dtype=np.float32)
+    for geno_sire in range(4):
+        for geno_dam in range(4):
+            for locus in range(n_loci):
+                output[geno_sire, geno_dam, locus] = (
+                    prob_sire[geno_sire, locus] * prob_dam[geno_dam, locus]
+                )
+    return output
+
+
+@jit(nopython=True, nogil=True)
+def create_child_segs(segregation_tensor, current_seg, output, n_loci):
+    """Creates the child-specific segregation tensor using the child's current segregation estimate.
+
+    :param segregation_tensor: the probability of each combination of
+        the sire's genotype, the dam's genotype, child's genotype and
+        segregation without any other information (P(p, m, allele, seg))
+    :type segregation_tensor: 4D numpy array of float32 with size 4 x 4 x 4 x 4
+    :param current_seg: The probability of each segregation of each locus of the child
+        with information from previous peeling cycle (P(seg))
+    :type current_seg: 2D numpy array of float32 with size 4 x n_loci
+    :param output: the probability of each combination of the sire's genotype,
+        the dam's genotype and the child's genotype of each locus with information
+        from previous peeling cycle (P(p, m, allele))
+    :type output: 4D numpy array of float32 with size 4 x 4 x 4 x n_loci
+    """
+    # Equivalent einsum: output = np.einsum("abcd, di -> abci", segregation_tensor, current_seg)
+    for geno_sire in range(4):
+        for geno_dam in range(4):
+            for geno_child in range(4):
+                for locus in range(n_loci):
+                    total = 0
+                    for seg in range(4):
+                        total += (
+                            segregation_tensor[geno_sire, geno_dam, geno_child, seg]
+                            * current_seg[seg, locus]
+                        )
+                    output[geno_sire, geno_dam, geno_child, locus] = total
+
+    return output
+
+
+@jit(nopython=True, nogil=True)
+def project_child_genotypes(child_segs, child_values, output, n_loci):
+    """Estimate the parental genotypes based on the child's genotypes and their segregation tensor.
+
+    :param child_segs: the probability of each combination of
+        the sire's genotype, the dam's genotype and the child's genotype of
+        each locus with information from previous peeling cycle
+        (P(p, m, allele))
+    :type child_segs: 4D numpy array of float32 with size 4 x 4 x 4 x n_loci
+    :param child_values: the probability of each genotype of each locus of the current child
+        given the information of itself and its later generations from previous peeling cycle
+        (P(allele))
+    :type child_values: 2D numpy array of float32 with size 4 x n_loci
+    :param output: the probability of each combination of the sire's genotype, the dam's genotype
+        given the information of later and current generations from previous peeling cycle
+        (P(p, m))
+    :type output: 3D numpy array of float32 with size 4 x 4 x n_loci
+    """
+    # Equivalent einsum: output = np.einsum("abci, ci -> abi", child_segs, child_values)
+    for geno_sire in range(4):
+        for geno_dam in range(4):
+            for locus in range(n_loci):
+                total = 0
+                for geno_child in range(4):
+                    total += (
+                        child_segs[geno_sire, geno_dam, geno_child, locus]
+                        * child_values[geno_child, locus]
+                    )
+                output[geno_sire, geno_dam, locus] = total
+
+    return output
+
+
+@jit(nopython=True, nogil=True)
+def project_parent_genotypes(child_segs, parent_values, output, n_loci):
+    """Project the parent genotypes down onto the child genotypes.
+
+    :param child_segs: the probability of each combination of
+        the sire's genotype, the dam's genotype and the child's genotype
+        of each locus with information from previous peeling cycle
+        (P(p, m, allele))
+    :type child_segs: 4D numpy array of float32 with size 4 x 4 x 4 x n_loci
+    :param parent_values: the probability of each combination of
+        sire's genotype and the dams's genotype of each locus given the
+        information of later and current generations from previous peeling cycle
+        without the information of the current child (P(p, m))
+    :type parent_values: 3D numpy array of float32 with size 4 x 4 x n_loci
+    :param output: the probability of child's genotype of each locus
+        given the later and current generations from previous peeling cycle
+        without the information of the current child (P(allele))
+    :type output: 2D numpy array of float32 with size 4 x n_loci
+    """
+    # Equivalent einsum: output = np.einsum("abci, abi -> ci", child_segs, parent_values)
+    for geno_child in range(4):
+        for locus in range(n_loci):
+            total = 0
+            for geno_sire in range(4):
+                for geno_dam in range(4):
+                    total += (
+                        child_segs[geno_sire, geno_dam, geno_child, locus]
+                        * parent_values[geno_sire, geno_dam, locus]
+                    )
+            output[geno_child, locus] = total
+
+    return output
+
+
+@jit(nopython=True, nogil=True)
+def add_log_child_to_parents(child_to_parents, all_to_parents, n_loci):
+    """Add one child's parent projection to the log-scale family total."""
+
+    for geno_sire in range(4):
+        for geno_dam in range(4):
+            for locus in range(n_loci):
+                all_to_parents[geno_sire, geno_dam, locus] += np.log(
+                    child_to_parents[geno_sire, geno_dam, locus]
+                )
+
+
+@jit(nopython=True, nogil=True)
+def add_log_child_to_parents_and_subtract_current(
+    child_to_parents, all_to_parents, parents_minus_current_child, n_loci
+):
+    """Add one child to the family total and store its negative log contribution."""
+
+    for geno_sire in range(4):
+        for geno_dam in range(4):
+            for locus in range(n_loci):
+                log_value = np.log(child_to_parents[geno_sire, geno_dam, locus])
+                all_to_parents[geno_sire, geno_dam, locus] += log_value
+                parents_minus_current_child[geno_sire, geno_dam, locus] = -log_value
+
+
+@jit(nopython=True, nogil=True)
+def add_joint_parents_and_all_to_minus(
+    parents_minus_child, joint_parents, all_to_parents, n_offspring, n_loci
+):
+    """Complete parent-minus-child log terms after all children are accumulated."""
+
+    for child_index in range(n_offspring):
+        for geno_sire in range(4):
+            for geno_dam in range(4):
+                for locus in range(n_loci):
+                    parents_minus_child[child_index, geno_sire, geno_dam, locus] += (
+                        np.log(joint_parents[geno_sire, geno_dam, locus])
+                        + all_to_parents[geno_sire, geno_dam, locus]
+                    )
+
+
+@jit(nopython=True, nogil=True)
+# pylint: disable=too-many-nested-blocks
+def estimate_segregation_with_norm(
+    segregation_tensors,
+    genotype_values,
+    output,
+    n_loci,
+):
+    """Estimate segregation probabilities with tensor-specific normalization.
+
+    :param segregation_tensor: the probability of each combination of
+        the sire's genotype, the dam's genotype and the child's genotype
+        and segregation without any other information (P(p, m, allele, seg))
+    :type segregation_tensor: 4D numpy array of float32 with size 4 x 4 x 4 x 4
+    :param segregation_tensor_norm: the mean probability of each combination of
+        the sire's genotype, the dam's genotype and the child's genotype across
+        the child's segregation without any other information
+        (1 / 4 x (P(p, m, allele)))
+    :type segregation_tensor_norm: 3D numpy array of float32 with size 4 x 4 x 4
+    :param parent_values: the probability of each combination of
+        sire's genotype and the dams's genotype of each locus given the
+        information of later and current generations from previous peeling cycle
+        without the information of the current child (P(p, m))
+    :type parent_values: 3D numpy array of float32 with size 4 x 4 x n_loci
+    :param child_values: the probability of each genotype of each locus of
+        the current child given the information of itself and its later
+        generations from previous peeling cycle (P(allele))
+    :type child_values: 2D numpy array of float32 with size 4 x n_loci
+    :param output: the probability of each segregation states of each
+        locus of the current child given the information of later and
+        current generations from previous peeling cycle (P(seg))
+    :type output: 2D numpy array of float32 with size 4 x n_loci
+    """
+    segregation_tensor, segregation_tensor_norm = segregation_tensors
+    parent_values, child_values = genotype_values
+    # Equivalent einsum before normalization: output =
+    # np.einsum("abcd, abi, ci -> di", segregation_tensor,
+    # parent_values, child_values)
+    for seg in range(4):
+        for locus in range(n_loci):
+            total = 0
+            for geno_sire in range(4):
+                for geno_dam in range(4):
+                    for geno_child in range(4):
+                        if (
+                            segregation_tensor_norm[geno_sire, geno_dam, geno_child]
+                            != 0
+                        ):
+                            total += (
+                                segregation_tensor[geno_sire, geno_dam, geno_child, seg]
+                                * parent_values[geno_sire, geno_dam, locus]
+                                * child_values[geno_child, locus]
+                                / segregation_tensor_norm[
+                                    geno_sire, geno_dam, geno_child
+                                ]
+                            )
+            output[seg, locus] = total
+    return output
+
+
+@jit(nopython=True, nogil=True)
+def combine_and_reduce_axis1(joint_estimate, parent_estimate, output, n_loci):
+    """Summing over axis 1 of joint_estimate with weights given by parent_estimate
+
+    :param joint_estimate: the probability of each combination of
+        sire's genotype and the dams's genotype of each locus
+        given the information of later and current generations
+        from previous peeling cycle (P(p, m))
+    :type joint_estimate: 3D numpy array of float32 with size 4 x 4 x n_loci
+    :param parent_estimate: the probability of each genotype of each locus of the dam
+        with information of the current child from previous peeling cycle (P(m))
+    :type parent_estimate: 2D numpy array of float32 with size 4 x n_loci
+    :return: the probability of each genotype of each locus of the sire
+        given the information of later and current generations from previous peeling cycle (P(p))
+    :rtype: 2D numpy array of float32 with size 4 x n_loci
+    """
+    # Equivalent einsum: output = np.einsum("abi, bi -> ai", joint_estimate, parent_estimate)
+    for geno_sire in range(4):
+        for locus in range(n_loci):
+            total = 0
+            for geno_dam in range(4):
+                total += (
+                    joint_estimate[geno_sire, geno_dam, locus]
+                    * parent_estimate[geno_dam, locus]
+                )
+            output[geno_sire, locus] = total
+    return output
+
+
+@jit(nopython=True, nogil=True)
+def combine_and_reduce_axis0(joint_estimate, parent_estimate, output, n_loci):
+    """Summing over axis 0 of joint_estimate with weights given by parent_estimate
+
+    :param joint_estimate: the probability of each combination of
+        sire's genotype and the dams's genotype of each locus
+        given the information of later and current generations
+        from previous peeling cycle (P(p, m))
+    :type joint_estimate: 3D numpy array of float32 with size 4 x 4 x n_loci
+    :param parent_estimate: the probability of each genotype of each locus of the sire
+        with information of the current child from previous peeling cycle (P(p))
+    :type parent_estimate: 2D numpy array of float32 with size 4 x n_loci
+    :return: the probability of each genotype of each locus of the dam
+        given the information of later and current generations from previous peeling cycle (P(m))
+    :rtype: 2D numpy array of float32 with size 4 x n_loci
+    """
+    # Equivalent einsum: output = np.einsum("abi, ai -> bi", joint_estimate, parent_estimate)
+    for geno_dam in range(4):
+        for locus in range(n_loci):
+            total = 0
+            for geno_sire in range(4):
+                total += (
+                    joint_estimate[geno_sire, geno_dam, locus]
+                    * parent_estimate[geno_sire, locus]
+                )
+            output[geno_dam, locus] = total
+    return output
+
+
+@jit(nopython=True, nogil=True)
+def normalize_4_by_locus(mat, n_loci):
+    """Normalize a 4 x n_loci probability matrix in place."""
+
+    for locus in range(n_loci):
+        total = 0
+        for a in range(4):
+            total += mat[a, locus]
+        for a in range(4):
+            mat[a, locus] /= total
+    return mat
+
+
+@jit(nopython=True, nogil=True)
+def normalize_16_by_locus(mat, n_loci):
+    """Normalize a 4 x 4 x n_loci probability tensor in place."""
+
+    for locus in range(n_loci):
+        total = 0
+        for a in range(4):
+            for b in range(4):
+                total += mat[a, b, locus]
+        for a in range(4):
+            for b in range(4):
+                mat[a, b, locus] /= total
+    return mat
+
+
+@jit(nopython=True, nogil=True)
+def smooth_4_by_locus(mat, scale, floor, n_loci):
+    """Apply a small in-place probability floor to a 4 x n_loci matrix."""
+
+    for a in range(4):
+        for locus in range(n_loci):
+            mat[a, locus] = mat[a, locus] * scale + floor
+    return mat
+
+
+@jit(nopython=True, nogil=True)
+def smooth_16_by_locus(mat, scale, floor, n_loci):
+    """Apply a small in-place probability floor to a 4 x 4 x n_loci tensor."""
+
+    for a in range(4):
+        for b in range(4):
+            for locus in range(n_loci):
+                mat[a, b, locus] = mat[a, b, locus] * scale + floor
+    return mat
+
+
+@jit(nopython=True, nogil=True)
+def exp_norm_2d(mat, n_loci):
+    """Output is to take the exponential of the matrix and normalize each locus.
+
+    :param mat: a 3D matrix with last axis represents the locus
+    :type mat: 3D numpy array of float32 with size 4 x 4 x n_loci
+    :return: the normalized exponential of the `mat`
+    :rtype: 3D numpy array of float32 with size 4 x 4 x n_loci
+    """
+    for locus in range(n_loci):
+        max_val = 1  # Log probabilities are non-positive, so 1 marks "unset".
+        for a in range(4):
+            for b in range(4):
+                if mat[a, b, locus] > max_val or max_val == 1:
+                    max_val = mat[a, b, locus]
+        for a in range(4):
+            for b in range(4):
+                mat[a, b, locus] -= max_val
+    tmp = np.exp(mat)
+    normalize_16_by_locus(tmp, n_loci)
+    return tmp
+
+
+@jit(nopython=True, nogil=True)
+def exp_norm_1d(mat, n_loci):
+    """Output is to take the exponential of the matrix and normalize each locus.
+
+    :param mat: a 2D matrix with last axis represents the locus
+    :type mat: 2D numpy array of float32 with size 4 x n_loci
+    :return: the normalized exponential of the `mat`
+    :rtype: 2D numpy array of float32 with size 4 x n_loci
+    """
+    for locus in range(n_loci):
+        max_val = 1  # Log probabilities are non-positive, so 1 marks "unset".
+        for a in range(4):
+            if mat[a, locus] > max_val or max_val == 1:
+                max_val = mat[a, locus]
+        for a in range(4):
+            mat[a, locus] -= max_val
+    tmp = np.exp(mat)
+    normalize_4_by_locus(tmp, n_loci)
+    return tmp
+
+
+@jit(nopython=True, nogil=True)
+def exp_norm_1d_in_place(mat, n_loci):
+    """Take the exponential of a 4 x n_loci log matrix and normalize in place."""
+
+    for locus in range(n_loci):
+        max_val = 1
+        for a in range(4):
+            if mat[a, locus] > max_val or max_val == 1:
+                max_val = mat[a, locus]
+        total = 0
+        for a in range(4):
+            mat[a, locus] = np.exp(mat[a, locus] - max_val)
+            total += mat[a, locus]
+        for a in range(4):
+            mat[a, locus] /= total
+    return mat
+
+
+@jit(nopython=True, nogil=True)
+def collapse_segregation_in_place(segregation, transmission, forward, n_loci):
+    """Using Baum-Welch algorithm to calculate the segregation probabilities.
+
+    :param segregation: the probability of each segregation states of
+        each locus of the current child given the information of later
+        and current generations from previous peeling cycle
+        (P(seg)). This is updated in place with the collapsed probabilities.
+        the segregation state ordering: pp, pm, mp, mm
+    :type segregation: 2D numpy array of float32 with size 4 x n_loci
+    :param transmission: transmission function based on the distance between adjacent loci
+    :type transmission: 1D numpy array of float32 with size (n_loci - 1)
+    :param forward: work array for forward messages.
+    :type forward: 2D numpy array of float32 with size 4 x n_loci
+    :return: None. The function updates segregation in place with the collapsed probabilities.
+    """
+    # Forward-backward pass over segregation states ordered as pp, pm, mp, mm.
+    tmp = np.full((4), 0, dtype=np.float32)
+    new = np.full((4), 0, dtype=np.float32)
+    prev = np.full((4), 1, dtype=np.float32)
+    buffers = (tmp, new, prev)
+
+    for state in range(4):
+        forward[state, 0] = 1
+
+    run_forward_segregation_pass(segregation, transmission, forward, buffers, n_loci)
+    run_backward_segregation_pass(segregation, transmission, forward, buffers, n_loci)
+    normalize_first_segregation_locus(segregation, forward, prev)
+
+
+@jit(
+    nopython=True,
+    nogil=True,
+    locals={"e": float32, "e2": float32, "e1e": float32, "e2i": float32},
+)
+def run_forward_segregation_pass(segregation, transmission, forward, buffers, n_loci):
+    """Run the forward pass for collapsed segregation probabilities."""
+
+    tmp, new, prev = buffers
+    for locus in range(1, n_loci):
+        e = transmission[locus - 1]
+        e2 = e**2
+        e1e = e * (1 - e)
+        e2i = (1.0 - e) ** 2
+        for state in range(4):
+            tmp[state] = prev[state] * segregation[state, locus - 1]
+
+        sum_state = 0
+        for state in range(4):
+            sum_state += tmp[state]
+        for state in range(4):
+            tmp[state] = tmp[state] / sum_state
+
+        new[0] = e2 * tmp[3] + e1e * (tmp[1] + tmp[2]) + e2i * tmp[0]
+        new[1] = e2 * tmp[2] + e1e * (tmp[0] + tmp[3]) + e2i * tmp[1]
+        new[2] = e2 * tmp[1] + e1e * (tmp[0] + tmp[3]) + e2i * tmp[2]
+        new[3] = e2 * tmp[0] + e1e * (tmp[1] + tmp[2]) + e2i * tmp[3]
+
+        for state in range(4):
+            forward[state, locus] = new[state]
+            prev[state] = new[state]
+
+
+@jit(
+    nopython=True,
+    nogil=True,
+    locals={"e": float32, "e2": float32, "e1e": float32, "e2i": float32},
+)
+def run_backward_segregation_pass(segregation, transmission, forward, buffers, n_loci):
+    """Run the backward pass for collapsed segregation probabilities."""
+
+    tmp, new, prev = buffers
+    for state in range(4):
+        prev[state] = 1
+
+    for locus in range(n_loci - 2, -1, -1):
+        e = transmission[locus]
+        e2 = e**2
+        e1e = e * (1 - e)
+        e2i = (1.0 - e) ** 2
+
+        for state in range(4):
+            tmp[state] = prev[state] * segregation[state, locus + 1]
+
+        sum_state = 0
+        for state in range(4):
+            sum_state += tmp[state]
+        for state in range(4):
+            tmp[state] = tmp[state] / sum_state
+
+        new[0] = e2 * tmp[3] + e1e * (tmp[1] + tmp[2]) + e2i * tmp[0]
+        new[1] = e2 * tmp[2] + e1e * (tmp[0] + tmp[3]) + e2i * tmp[1]
+        new[2] = e2 * tmp[1] + e1e * (tmp[0] + tmp[3]) + e2i * tmp[2]
+        new[3] = e2 * tmp[0] + e1e * (tmp[1] + tmp[2]) + e2i * tmp[3]
+
+        sum_state = 0
+        for state in range(4):
+            segregation[state, locus + 1] = (
+                segregation[state, locus + 1] * forward[state, locus + 1] * prev[state]
+            )
+            sum_state += segregation[state, locus + 1]
+        for state in range(4):
+            segregation[state, locus + 1] = segregation[state, locus + 1] / sum_state
+            prev[state] = new[state]
+
+
+@jit(nopython=True, nogil=True)
+def normalize_first_segregation_locus(segregation, forward, prev):
+    """Normalize the first locus after the backward segregation pass."""
+
+    sum_state = 0
+    for state in range(4):
+        segregation[state, 0] = segregation[state, 0] * forward[state, 0] * prev[state]
+        sum_state += segregation[state, 0]
+    for state in range(4):
+        segregation[state, 0] = segregation[state, 0] / sum_state
